requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and authentication
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP and file handling
httpx>=0.24.0,<0.25.0
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0

# Production monitoring (optional)
# sentry-sdk[fastapi]==1.38.0


